+++ CHANGELOG.md
@@ -1,5 +1,3 @@
 ### Changes ###
 
-  * Fix chat install
-  * Created an animated Resting Indicator
-  * Added Skin for 'I Remember You'
+  * Fix skin issues if the desired AddOn is not enabled